add view by email

+++ apis.py
@@ -2,6 +2,7 @@ import time
 from datetime import datetime
 from flask import request, jsonify, Blueprint, send_from_directory
 import os
+from db import add_user_uploads,get_records_by_email
 from google_cloud import upload_file,upload_to_gemini
 from helpers import allowed_file
 
@@ -19,22 +20,46 @@ def upload():
     if not allowed_file(image.filename):
         return jsonify({"error": "Uploaded file is not an allowed image type"}), 400
 
+    user_email = request.form.get('email')
+    if not user_email:
+        return jsonify({"error": "No email provided"}), 400
+
     timestamp = datetime.now().strftime('%Y%m%d%H%M%S')
-    local_path = timestamp+'_'+image.filename
+    local_path = timestamp + '_' + image.filename
     image.save(local_path)
     image.seek(0)
-    public_url = upload_file(image, "image_files/"+local_path)
-    text=upload_to_gemini(local_path,timestamp)
+    public_url = upload_file(image, "image_files/" + local_path)
+    text, txt_url = upload_to_gemini(local_path, timestamp)
     os.remove(local_path)
-    return jsonify({"message": "Image uploaded successfully ", "Transcription": text, "url": public_url}), 200
+
+    data = {
+        'email':user_email,
+        'image_path':public_url,
+        'transcript':text,
+        'text_path':txt_url
+    }
+    add_user_uploads(data)
+    return jsonify({"result": data}), 200
+
+
+@upload_api.route('/uploads', methods=['GET'])
+def get_uploads_by_email():
+    user_email = request.args.get('email')
+
+    if not user_email:
+        return jsonify({"error": "No email provided"}), 400
+
+    records = get_records_by_email(user_email)
+
+    if not records:
+        return jsonify({"error": "No records found for this email"}), 404
+
+    return jsonify({"uploads": records}), 200
 
 
 @health_api.route('/health', methods=['GET'])
 def health():
     time.sleep(5)
-    if not os.path.exists("tmp"):
-        logger.info("Creating tmp directory")
-
     return jsonify({"message": "Server Health : Running"}), 200
 
 @ui_api.route('/')
